add select scrollbar, will diaplay error message when edge file is invalid
