+++ src/index.js
@@ -4,6 +4,7 @@ import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import "@blueprintjs/core/lib/css/blueprint.css";
 import "@blueprintjs/table/lib/css/table.css";
+import "@blueprintjs/select/lib/css/blueprint-select.css";
 import "normalize.css";
 
 ReactDOM.render(
